requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
jinja2==3.1.2
openai==1.2.4
httpx==0.25.1 